+++ src/browser/connection/error-hints.ts
@@ -1,0 +1,7 @@
+enum BrowserConnectionErrorHint {
+    TooHighConcurrencyFactor = 'TooHighConcurrencyFactor',
+    UseBrowserInitOption = 'UseBrowserInitOption',
+    RestErrorCauses = 'RestErrorCauses'
+}
+
+export default BrowserConnectionErrorHint;