+++ src/errors/types.js
@@ -146,5 +146,8 @@ export const RUNTIME_ERRORS = {
     uncaughtErrorInReporter:                            'E1058',
     roleInitializedWithRelativeUrl:                     'E1059',
     typeScriptCompilerLoadingError:                     'E1060',
-    cannotCustomizeSpecifiedCompilers:                  'E1061'
+    cannotCustomizeSpecifiedCompilers:                  'E1061',
+    cannotEnableRetryTestPagesOption:                   'E1062',
+    browserConnectionError:                             'E1063',
+    testRunRequestInDisconnectedBrowser:                'E1064'
 };