+++ test/functional/fixtures/regression/gh-5239/test.js
@@ -1,12 +1,22 @@
 const http           = require('http');
 const path           = require('path');
+const expect         = require('chai').expect;
 const config         = require('../../../config');
 const createTestCafe = require('../../../../../lib');
+
+const { createReporter } = require('../../../utils/reporter');
 
 const ERROR_RESPONSE_COUNT        = 8;
 const SIGNIFICANT_REQUEST_TIMEOUT = 200;
 
 let previousRequestTime = null;
+let warnings            = [];
+
+const customReporter = createReporter({
+    async reportTaskDone (endTime, passed, warns) {
+        warnings = warns;
+    }
+});
 
 function createServer () {
     let requestCounter = 0;
@@ -35,31 +45,51 @@ function createServer () {
     return server;
 }
 
-async function run () {
-    const testcafe = await createTestCafe('localhost', 1335, 1336, void 0, true, true);
+async function run ({ src, browsers, retryTestPages, reporter }) {
+    const testcafe = await createTestCafe('localhost', 1335, 1336, void 0, true, retryTestPages);
     const runner   = testcafe.createRunner();
 
     await runner
-        .src(path.join(__dirname, './testcafe-fixtures/index.js'))
-        .browsers('chrome --headless')
-        .run();
+        .src(path.join(__dirname, src))
+        .browsers(browsers);
+
+    if (reporter)
+        runner.reporter(reporter);
+
+    await runner.run();
 
     await testcafe.close();
 }
 
 const isLocalChrome = config.useLocalBrowsers && config.browsers.some(browser => browser.alias.indexOf('chrome') > -1);
 
-if (isLocalChrome) {
-    describe('[Regression](GH-5239)', function () {
+describe('[Regression](GH-5239)', function () {
+    if (isLocalChrome) {
         it('Should make multiple request for the page if the server does not respond', function () {
             this.timeout(30000);
 
             const server = createServer();
 
-            return run()
+            return run({ retryTestPages: true, browsers: 'chrome --headless', src: './testcafe-fixtures/index.js' })
                 .then(() => {
                     return server.close();
                 });
         });
-    });
-}
+    }
+
+    if (config.currentEnvironmentName === config.testingEnvironmentNames.localBrowsersIE) {
+        it('Should show warning if the \'retryTestPages\' option is not supported', function () {
+            return run({
+                retryTestPages: true,
+                browsers:       'ie',
+                src:            './testcafe-fixtures/warnings-test.js',
+                reporter:       customReporter
+            })
+                .then(() => {
+                    expect(warnings).eql([
+                        'Cannot enable the \'retryTestPages\' option in "ie". Please ensure that your version of "ie" supports the Service Worker API (https://developer.mozilla.org/en-US/docs/Web/API/Service_Worker_API).\n'
+                    ]);
+                });
+        });
+    }
+});