+++ ts-defs-src/runner-api/runner-api.d.ts
@@ -320,6 +320,10 @@ interface RunOptions {
      */
     pageLoadTimeout: number;
     /**
+     * Specifies the time (in milliseconds) TestCafe waits for the browser to start
+     */
+    browserInitTimeout: number;
+    /**
      * Specifies the test execution speed. A number between 1 (fastest) and 0.01 (slowest). If an individual action's speed is also specified, the action speed setting overrides the test speed.
      */
     speed: number;
@@ -335,6 +339,14 @@ interface RunOptions {
      * Defines whether to disable page caching during test execution.
      */
     disablePageCaching: boolean;
+    /**
+     * Specifies the timeout in milliseconds to complete the request for the page's HTML
+     */
+    pageRequestTimeout: number;
+    /**
+     * Specifies the timeout in milliseconds to complete the AJAX requests (XHR or fetch)
+     */
+    ajaxRequestTimeout: number;
 }
 
 interface TestCafeFactory {
@@ -343,6 +355,7 @@ interface TestCafeFactory {
         port1?: number,
         port2?: number,
         sslOptions?: TlsOptions,
-        developmentMode?: boolean
+        developmentMode?: boolean,
+        retryTestPages?: boolean
     ): Promise<TestCafe>;
 }